+++ ldbc_socialnet_dbgen/src/main/java/ldbc/socialnet/dbgen/objects/Photo.java
@@ -43,13 +43,13 @@ public class Photo {
     String image;
     long albumId; 
     int locationIdx; 
-    int creatorId;		// Id of user's account
+    long creatorId;		// Id of user's account
     String locationName;
     double latt; 
     double longt; 
     long takenTime; 
     TreeSet<Integer> tags;
-    int[] interestedUserAccs;
+    long[] interestedUserAccs;
     long[] interestedUserAccsTimestamp;
 
     IP ipAddress; 
@@ -57,10 +57,10 @@ public class Photo {
 
     byte 	browserIdx; 
 
-    public int getCreatorId() {
+    public long getCreatorId() {
         return creatorId;
     }
-    public void setCreatorId(int creatorId) {
+    public void setCreatorId(long creatorId) {
         this.creatorId = creatorId;
     }
     public byte getBrowserIdx() {
@@ -138,10 +138,10 @@ public class Photo {
     public void setLocationName(String locationName) {
         this.locationName = locationName;
     }
-    public int[] getInterestedUserAccs() {
+    public long[] getInterestedUserAccs() {
         return interestedUserAccs;
     }
-    public void setInterestedUserAccs(int[] interestedUserAccs) {
+    public void setInterestedUserAccs(long[] interestedUserAccs) {
         this.interestedUserAccs = interestedUserAccs;
     }	
     public long[] getInterestedUserAccsTimestamp() {
